requirements.txt
# Hugging Face + Diffusers
diffusers==0.35.2
transformers==4.46.2
accelerate==0.26.0
safetensors==0.4.3

#API
fastapi==0.101.0
uvicorn==0.23.2

# Media & Utils
Pillow==10.4.0
imageio[ffmpeg]==2.31.1
numpy==1.26.4